Move evidence and URL aliases to one mixin
Documentable mixin is misused and can not be contain aliases for those
two fields so until that mixin is refactored and fixed we will have a
temporary one with proper definitions for evidence and URLs.
