+++ src/ggrc/models/request.py
@@ -26,13 +26,13 @@ from ggrc.models.mixins.with_similarity_score import WithSimilarityScore
 from ggrc.models.mixins.autostatuschangeable import AutoStatusChangeable
 from ggrc.models.deferred import deferred
 from ggrc.models.mixins.assignable import Assignable
-from ggrc.models.object_document import Documentable
+from ggrc.models.object_document import EvidenceURL
 from ggrc.models.object_person import Personable
 from ggrc.utils import similarity_options as similarity_options_module
 
 
 class Request(statusable.Statusable, AutoStatusChangeable, Assignable,
-              Documentable, Personable, CustomAttributable,
+              EvidenceURL, Personable, CustomAttributable,
               relationship.Relatable, WithSimilarityScore, Titled, Slugged,
               Described, Commentable, FinishedDate, VerifiedDate, Base,
               db.Model):
@@ -141,16 +141,6 @@ class Request(statusable.Statusable, AutoStatusChangeable, Assignable,
           "filter_by": "_filter_by_related_verifiers",
           "type": reflection.AttributeInfo.Type.MAPPING,
       },
-      "document_url": {
-          "display_name": "Url",
-          "filter_by": "_filter_by_url",
-          "type": reflection.AttributeInfo.Type.SPECIAL_MAPPING,
-      },
-      "document_evidence": {
-          "display_name": "Evidence",
-          "filter_by": "_filter_by_evidence",
-          "type": reflection.AttributeInfo.Type.SPECIAL_MAPPING,
-      },
   }
 
   def _display_name(self):
@@ -186,14 +176,6 @@ class Request(statusable.Statusable, AutoStatusChangeable, Assignable,
   @classmethod
   def _filter_by_related_verifiers(cls, predicate):
     return cls._get_relate_filter(predicate, "Verifier")
-
-  @classmethod
-  def _filter_by_url(cls, _):
-    return None
-
-  @classmethod
-  def _filter_by_evidence(cls, _):
-    return None
 
   @classmethod
   def _filter_by_request_audit(cls, predicate):
